Always ensure that yq is correct

+++ Next_Generation_Node_B/generate_configurations.sh
@@ -80,6 +80,12 @@ fi
 if ! command -v yq &>/dev/null; then
     sudo "$SCRIPT_DIR/install_scripts/./install_yq.sh"
 fi
+# Check that the correct version of yq is installed
+if ! yq --version 2>/dev/null | grep -q 'https://github\.com/mikefarah/yq'; then
+    echo "ERROR: Detected an incompatible yq installation."
+    echo "Please ensure the Python yq is uninstalled with \"pip uninstall yq\", then re-run this script."
+    exit 1
+fi
 
 echo "Restoring gNodeB configuration file..."
 rm -rf configs
